devops/l8/microservice2/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
